backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.35.0
pandas==2.3.1
openpyxl==3.1.5
pypdf==5.9.0
openai==2.11.0
pydantic==2.11.7
python-dotenv==1.1.1
python-multipart==0.0.20
pytest==8.4.1

